Arrange directories
