+++ crates/oxc_prettier/src/format/js.rs
@@ -1,0 +1,1757 @@
+use std::borrow::Cow;
+
+use cow_utils::CowUtils;
+use oxc_allocator::Vec;
+use oxc_ast::{ast::*, AstKind};
+use oxc_span::GetSpan;
+use oxc_syntax::identifier::{is_identifier_name, is_line_terminator};
+
+use crate::{
+    array, dynamic_text,
+    format::{
+        print::{
+            array, arrow_function, assignment, binaryish, block, call_expression, class, function,
+            function_parameters, misc, module, object, property, string, template_literal, ternary,
+        },
+        Format,
+    },
+    group, hardline, indent,
+    ir::{Doc, JoinSeparator},
+    join, line, softline, text, wrap, Prettier,
+};
+
+impl<'a> Format<'a> for Program<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, Program, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            // In Prettier, this is treated as a comment
+            if let Some(hashbang) = &self.hashbang {
+                parts.push(hashbang.format(p));
+            }
+
+            if let Some(body_doc) = block::print_block_body(p, &self.body, Some(&self.directives)) {
+                parts.push(body_doc);
+                // XXX: Prettier seems to add this, but test results don't match
+                // parts.extend(hardline!());
+            }
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for Hashbang<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(dynamic_text!(p, self.span.source_text(p.source_text)));
+        parts.extend(hardline!());
+        // Preserve original newline
+        if let Some(c) = p.source_text[self.span.end as usize..].chars().nth(1) {
+            if is_line_terminator(c) {
+                parts.extend(hardline!());
+            }
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for Directive<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(dynamic_text!(
+            p,
+            string::print_string(p, self.directive.as_str(), p.options.single_quote,)
+        ));
+        if let Some(semi) = p.semi() {
+            parts.push(semi);
+        }
+        parts.extend(hardline!());
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for Statement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::BlockStatement(stmt) => stmt.format(p),
+            Self::BreakStatement(stmt) => stmt.format(p),
+            Self::ContinueStatement(stmt) => stmt.format(p),
+            Self::DebuggerStatement(stmt) => stmt.format(p),
+            Self::DoWhileStatement(stmt) => stmt.format(p),
+            Self::EmptyStatement(stmt) => stmt.format(p),
+            Self::ExpressionStatement(stmt) => stmt.format(p),
+            Self::ForInStatement(stmt) => stmt.format(p),
+            Self::ForOfStatement(stmt) => stmt.format(p),
+            Self::ForStatement(stmt) => stmt.format(p),
+            Self::IfStatement(stmt) => stmt.format(p),
+            Self::LabeledStatement(stmt) => stmt.format(p),
+            Self::ReturnStatement(stmt) => stmt.format(p),
+            Self::SwitchStatement(stmt) => stmt.format(p),
+            Self::ThrowStatement(stmt) => stmt.format(p),
+            Self::TryStatement(stmt) => stmt.format(p),
+            Self::WhileStatement(stmt) => stmt.format(p),
+            Self::WithStatement(stmt) => stmt.format(p),
+            match_module_declaration!(Self) => self.to_module_declaration().format(p),
+            match_declaration!(Self) => self.to_declaration().format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ExpressionStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ExpressionStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(self.expression.format(p));
+            if let Some(semi) = p.semi() {
+                parts.push(semi);
+            }
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for EmptyStatement {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        text!("")
+    }
+}
+
+impl<'a> Format<'a> for IfStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, IfStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            let test_doc = self.test.format(p);
+            let consequent = self.consequent.format(p);
+            let consequent = misc::adjust_clause(p, &self.consequent, consequent, false);
+
+            let opening = group!(
+                p,
+                [
+                    text!("if ("),
+                    group!(p, [indent!(p, [softline!(), test_doc]), softline!(),]),
+                    text!(")"),
+                    consequent
+                ]
+            );
+            parts.push(opening);
+
+            if let Some(alternate) = &self.alternate {
+                let else_on_same_line = matches!(alternate, Statement::BlockStatement(_));
+                if else_on_same_line {
+                    parts.push(text!(" "));
+                } else {
+                    parts.extend(hardline!());
+                }
+                parts.push(text!("else"));
+                let alternate_doc = alternate.format(p);
+                parts.push(group!(
+                    p,
+                    [misc::adjust_clause(
+                        p,
+                        alternate,
+                        alternate_doc,
+                        matches!(alternate, Statement::IfStatement(_)),
+                    )]
+                ));
+            }
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for BlockStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, BlockStatement, { block::print_block(p, &self.body, None) })
+    }
+}
+
+impl<'a> Format<'a> for ForStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ForStatement, {
+            let body = self.body.format(p);
+            let body = misc::adjust_clause(p, &self.body, body, false);
+
+            if self.init.is_none() && self.test.is_none() && self.update.is_none() {
+                return group!(p, [text!("for (;;)"), body]);
+            }
+
+            let parts_head = {
+                let mut parts_head = Vec::new_in(p.allocator);
+                parts_head.push(softline!());
+                if let Some(init) = &self.init {
+                    parts_head.push(init.format(p));
+                }
+                parts_head.push(text!(";"));
+                parts_head.push(line!());
+                if let Some(init) = &self.test {
+                    parts_head.push(init.format(p));
+                }
+                parts_head.push(text!(";"));
+                parts_head.push(line!());
+                if let Some(init) = &self.update {
+                    parts_head.push(init.format(p));
+                }
+                indent!(p, parts_head)
+            };
+
+            group!(p, [text!("for ("), group!(p, [parts_head, softline!()]), text!(")"), body,])
+        })
+    }
+}
+
+impl<'a> Format<'a> for ForStatementInit<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            ForStatementInit::VariableDeclaration(v) => v.format(p),
+            match_expression!(ForStatementInit) => self.to_expression().format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ForInStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ForInStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("for ("));
+            parts.push(self.left.format(p));
+            parts.push(text!(" in "));
+            parts.push(self.right.format(p));
+            parts.push(text!(")"));
+            let body = self.body.format(p);
+            parts.push(misc::adjust_clause(p, &self.body, body, false));
+            group!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for ForOfStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ForOfStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("for"));
+            if self.r#await {
+                parts.push(text!(" await"));
+            }
+            parts.push(text!(" ("));
+            parts.push(self.left.format(p));
+            parts.push(text!(" of "));
+            parts.push(self.right.format(p));
+            parts.push(text!(")"));
+            let body = self.body.format(p);
+            parts.push(misc::adjust_clause(p, &self.body, body, false));
+            group!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for ForStatementLeft<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            ForStatementLeft::VariableDeclaration(v) => v.format(p),
+            match_assignment_target!(ForStatementLeft) => self.to_assignment_target().format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for WhileStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, WhileStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            parts.push(text!("while ("));
+            let test_doc = self.test.format(p);
+            parts.push(group!(p, [indent!(p, [softline!(), test_doc]), softline!(),]));
+            parts.push(text!(")"));
+
+            let body = self.body.format(p);
+            parts.push(misc::adjust_clause(p, &self.body, body, false));
+
+            group!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for DoWhileStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, DoWhileStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            let clause = self.body.format(p);
+            let clause = misc::adjust_clause(p, &self.body, clause, false);
+            let do_body = group!(p, [text!("do"), clause]);
+
+            parts.push(do_body);
+
+            if matches!(self.body, Statement::BlockStatement(_)) {
+                parts.push(text!(" "));
+            } else {
+                parts.extend(hardline!());
+            }
+
+            parts.push(text!("while ("));
+            let test_doc = self.test.format(p);
+            parts.push(group!(p, [indent!(p, [softline!(), test_doc]), softline!()]));
+            parts.push(text!(")"));
+            if let Some(semi) = p.semi() {
+                parts.push(semi);
+            }
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for ContinueStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("continue"));
+
+        if let Some(label) = &self.label {
+            parts.push(text!(" "));
+            parts.push(label.format(p));
+        }
+
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for BreakStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("break"));
+
+        if let Some(label) = &self.label {
+            parts.push(text!(" "));
+            parts.push(label.format(p));
+        }
+
+        if p.options.semi {
+            parts.push(text!(";"));
+        }
+
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for SwitchStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, SwitchStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            let mut header_parts = Vec::new_in(p.allocator);
+
+            header_parts.push(text!("switch ("));
+
+            let discriminant_doc = self.discriminant.format(p);
+            header_parts.push(indent!(p, [softline!(), discriminant_doc]));
+
+            header_parts.push(softline!());
+            header_parts.push(text!(")"));
+
+            parts.push(group!(p, header_parts));
+
+            parts.push(text!(" {"));
+
+            let mut cases_parts = Vec::new_in(p.allocator);
+            let len = self.cases.len();
+            for (i, case) in self.cases.iter().enumerate() {
+                cases_parts.push({
+                    let mut parts = Vec::new_in(p.allocator);
+                    parts.extend(hardline!());
+                    parts.push(case.format(p));
+                    indent!(p, parts)
+                });
+                if i != len - 1 && p.is_next_line_empty(case.span) {
+                    cases_parts.extend(hardline!());
+                }
+            }
+            parts.extend(cases_parts);
+
+            parts.extend(hardline!());
+            parts.push(text!("}"));
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for SwitchCase<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+
+        if let Some(test) = &self.test {
+            parts.push(text!("case "));
+            parts.push(test.format(p));
+            parts.push(text!(":"));
+        } else {
+            parts.push(text!("default:"));
+        }
+
+        let consequent: Vec<_> = Vec::from_iter_in(
+            self.consequent.iter().filter(|c| !matches!(c, Statement::EmptyStatement(_))),
+            p.allocator,
+        );
+        let len = consequent.len();
+        let is_only_one_block_statement =
+            len == 1 && matches!(self.consequent[0], Statement::BlockStatement(_));
+
+        let mut consequent_parts = Vec::new_in(p.allocator);
+        for i in 0..len {
+            let stmt = &consequent[i];
+
+            if i != 0 && matches!(stmt, Statement::BreakStatement(_)) {
+                let last_stmt = &consequent[i - 1];
+                if p.is_next_line_empty(last_stmt.span()) {
+                    consequent_parts.extend(hardline!());
+                }
+            }
+
+            if is_only_one_block_statement {
+                consequent_parts.push(text!(" "));
+            } else {
+                consequent_parts.extend(hardline!());
+            }
+            consequent_parts.push(stmt.format(p));
+        }
+
+        if !consequent_parts.is_empty() {
+            if is_only_one_block_statement {
+                parts.extend(consequent_parts);
+            } else {
+                parts.push(indent!(p, [group!(p, consequent_parts)]));
+            }
+        }
+
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ReturnStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ReturnStatement, {
+            array!(
+                p,
+                [
+                    text!("return"),
+                    function::print_return_or_throw_argument(p, self.argument.as_ref())
+                ]
+            )
+        })
+    }
+}
+
+impl<'a> Format<'a> for LabeledStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        if matches!(self.body, Statement::EmptyStatement(_)) {
+            let label_doc = self.label.format(p);
+            return array!(p, [label_doc, text!(":;")]);
+        }
+
+        let label_doc = self.label.format(p);
+        let body_doc = self.body.format(p);
+        array!(p, [label_doc, text!(": "), body_doc])
+    }
+}
+
+impl<'a> Format<'a> for TryStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, TryStatement, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            parts.push(text!("try "));
+            parts.push(self.block.format(p));
+            if let Some(handler) = &self.handler {
+                parts.push(text!(" "));
+                parts.push(handler.format(p));
+            }
+            if let Some(finalizer) = &self.finalizer {
+                parts.push(text!(" finally "));
+                parts.push(finalizer.format(p));
+            }
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for CatchClause<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, CatchClause, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            parts.push(text!("catch "));
+            if let Some(param) = &self.param {
+                parts.push(text!("("));
+                parts.push(param.pattern.format(p));
+                parts.push(text!(") "));
+            }
+            parts.push(self.body.format(p));
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for ThrowStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ThrowStatement, {
+            array!(
+                p,
+                [text!("throw"), function::print_return_or_throw_argument(p, Some(&self.argument))]
+            )
+        })
+    }
+}
+
+impl<'a> Format<'a> for WithStatement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let object_doc = self.object.format(p);
+        let body_doc = self.body.format(p);
+        group!(
+            p,
+            [
+                text!("with ("),
+                object_doc,
+                text!(")"),
+                misc::adjust_clause(p, &self.body, body_doc, false)
+            ]
+        )
+    }
+}
+
+impl<'a> Format<'a> for DebuggerStatement {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("debugger"));
+
+        if p.options.semi {
+            parts.push(text!(";"));
+        }
+
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ModuleDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ModuleDeclaration, {
+            if let ModuleDeclaration::ImportDeclaration(decl) = self {
+                decl.format(p)
+            } else {
+                module::print_export_declaration(p, self)
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for Declaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::VariableDeclaration(stmt) => stmt.format(p),
+            Self::FunctionDeclaration(stmt) => stmt.format(p),
+            Self::ClassDeclaration(decl) => decl.format(p),
+            Self::TSTypeAliasDeclaration(decl) => decl.format(p),
+            Self::TSInterfaceDeclaration(decl) => decl.format(p),
+            Self::TSEnumDeclaration(decl) => decl.format(p),
+            Self::TSModuleDeclaration(decl) => decl.format(p),
+            Self::TSImportEqualsDeclaration(decl) => decl.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for VariableDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, VariableDeclaration, {
+            // We generally want to terminate all variable declarations with a
+            // semicolon, except when they in the () part of for loops.
+            let parent_for_loop = match p.parent_kind() {
+                AstKind::ForStatement(stmt) => Some(stmt.body.span()),
+                AstKind::ForInStatement(stmt) => Some(stmt.body.span()),
+                AstKind::ForOfStatement(stmt) => Some(stmt.body.span()),
+                _ => None,
+            };
+
+            let kind = self.kind.as_str();
+
+            let mut parts = Vec::new_in(p.allocator);
+
+            if self.declare {
+                parts.push(text!("declare "));
+            }
+
+            parts.push(text!(kind));
+            parts.push(text!(" "));
+
+            let is_hardline = !p.parent_kind().is_iteration_statement()
+                && self.declarations.iter().all(|decl| decl.init.is_some());
+            let decls_len = self.declarations.len();
+            parts.extend(self.declarations.iter().enumerate().map(|(i, decl)| {
+                if decls_len > 1 {
+                    let mut d_parts = Vec::new_in(p.allocator);
+                    if i != 0 {
+                        d_parts.push(text!(","));
+                        if is_hardline {
+                            d_parts.extend(hardline!());
+                        } else {
+                            d_parts.push(line!());
+                        }
+                    }
+                    d_parts.push(decl.format(p));
+                    indent!(p, d_parts)
+                } else {
+                    decl.format(p)
+                }
+            }));
+
+            if !parent_for_loop.is_some_and(|span| span != self.span) {
+                if let Some(semi) = p.semi() {
+                    parts.push(semi);
+                }
+            }
+
+            group!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for VariableDeclarator<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, VariableDeclarator, { assignment::print_variable_declarator(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for Function<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, Function, { function::print_function(p, self, None) })
+    }
+}
+
+impl<'a> Format<'a> for FunctionBody<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, FunctionBody, {
+            block::print_block(p, &self.statements, Some(&self.directives))
+        })
+    }
+}
+
+impl<'a> Format<'a> for FormalParameters<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, FormalParameters, {
+            function_parameters::print_function_parameters(p, self)
+        })
+    }
+}
+
+impl<'a> Format<'a> for FormalParameter<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, FormalParameter, { self.pattern.format(p) })
+    }
+}
+
+impl<'a> Format<'a> for ImportDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("import"));
+        if self.import_kind.is_type() {
+            parts.push(text!(" type"));
+        }
+
+        if let Some(specifiers) = &self.specifiers {
+            let is_default = specifiers.first().is_some_and(|x| {
+                matches!(x, ImportDeclarationSpecifier::ImportDefaultSpecifier(_))
+            });
+
+            let validate_namespace = |x: &ImportDeclarationSpecifier| {
+                matches!(x, ImportDeclarationSpecifier::ImportNamespaceSpecifier(_))
+            };
+
+            let is_namespace = specifiers.first().is_some_and(validate_namespace)
+                || specifiers.get(1).is_some_and(validate_namespace);
+
+            parts.push(module::print_module_specifiers(p, specifiers, is_default, is_namespace));
+            parts.push(text!(" from"));
+        }
+        parts.push(text!(" "));
+        parts.push(self.source.format(p));
+
+        if let Some(with_clause) = &self.with_clause {
+            parts.push(text!(" "));
+            parts.push(with_clause.format(p));
+        }
+
+        if let Some(semi) = p.semi() {
+            parts.push(semi);
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ImportDeclarationSpecifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::ImportSpecifier(specifier) => specifier.format(p),
+            Self::ImportDefaultSpecifier(specifier) => specifier.format(p),
+            Self::ImportNamespaceSpecifier(specifier) => specifier.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ImportSpecifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let typed = if self.import_kind.is_type() { text!("type ") } else { text!("") };
+
+        if self.imported.span() == self.local.span {
+            let local_doc = self.local.format(p);
+            array!(p, [typed, local_doc])
+        } else {
+            let imported_doc = self.imported.format(p);
+            let local_doc = self.local.format(p);
+            array!(p, [typed, imported_doc, text!(" as "), local_doc])
+        }
+    }
+}
+
+impl<'a> Format<'a> for ImportDefaultSpecifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        self.local.format(p)
+    }
+}
+
+impl<'a> Format<'a> for ImportNamespaceSpecifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let local_doc = self.local.format(p);
+        array!(p, [text!("* as "), local_doc])
+    }
+}
+
+impl<'a> Format<'a> for WithClause<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let attribute_keyword_doc = self.attributes_keyword.format(p);
+        let with_clause_doc = object::print_object(p, object::ObjectLike::WithClause(self));
+        array!(p, [attribute_keyword_doc, text!(" "), with_clause_doc])
+    }
+}
+
+impl<'a> Format<'a> for ImportAttribute<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let key_doc = self.key.format(p);
+        let value_doc = self.value.format(p);
+        array!(p, [key_doc, text!(": "), value_doc])
+    }
+}
+
+impl<'a> Format<'a> for ImportAttributeKey<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::Identifier(ident) => ident.format(p),
+            Self::StringLiteral(literal) => literal.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ExportNamedDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        if let Some(decl) = &self.declaration {
+            parts.push(text!(" "));
+            parts.push(decl.format(p));
+        } else {
+            parts.push(module::print_module_specifiers(
+                p,
+                &self.specifiers,
+                /* include_default */ false,
+                /* include_namespace */ false,
+            ));
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ExportSpecifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        if self.exported.span() == self.local.span() {
+            self.local.format(p)
+        } else {
+            let local_doc = self.local.format(p);
+            let exported_doc = self.exported.format(p);
+            array!(p, [local_doc, text!(" as "), exported_doc])
+        }
+    }
+}
+
+impl<'a> Format<'a> for ModuleExportName<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::IdentifierName(ident) => ident.format(p),
+            Self::IdentifierReference(ident) => ident.format(p),
+            Self::StringLiteral(literal) => literal.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ExportAllDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!(" *"));
+        if let Some(exported) = &self.exported {
+            parts.push(text!(" as "));
+            parts.push(exported.format(p));
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ExportDefaultDeclaration<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        self.declaration.format(p)
+    }
+}
+impl<'a> Format<'a> for ExportDefaultDeclarationKind<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            match_expression!(Self) => self.to_expression().format(p),
+            Self::FunctionDeclaration(decl) => decl.format(p),
+            Self::ClassDeclaration(decl) => decl.format(p),
+            Self::TSInterfaceDeclaration(decl) => decl.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for Expression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::BooleanLiteral(lit) => lit.format(p),
+            Self::NullLiteral(lit) => lit.format(p),
+            Self::NumericLiteral(lit) => lit.format(p),
+            Self::BigIntLiteral(lit) => lit.format(p),
+            Self::RegExpLiteral(lit) => lit.format(p),
+            Self::StringLiteral(lit) => lit.format(p),
+            Self::Identifier(ident) => ident.format(p),
+            Self::ThisExpression(expr) => expr.format(p),
+            match_member_expression!(Self) => self.to_member_expression().format(p),
+            Self::CallExpression(expr) => expr.format(p),
+            Self::ArrayExpression(expr) => expr.format(p),
+            Self::ObjectExpression(expr) => expr.format(p),
+            Self::FunctionExpression(expr) => expr.format(p),
+            Self::ArrowFunctionExpression(expr) => expr.format(p),
+            Self::YieldExpression(expr) => expr.format(p),
+            Self::UpdateExpression(expr) => expr.format(p),
+            Self::UnaryExpression(expr) => expr.format(p),
+            Self::BinaryExpression(expr) => expr.format(p),
+            Self::PrivateInExpression(expr) => expr.format(p),
+            Self::LogicalExpression(expr) => expr.format(p),
+            Self::ConditionalExpression(expr) => expr.format(p),
+            Self::AssignmentExpression(expr) => expr.format(p),
+            Self::SequenceExpression(expr) => expr.format(p),
+            Self::ParenthesizedExpression(expr) => expr.format(p),
+            Self::ImportExpression(expr) => expr.format(p),
+            Self::TemplateLiteral(literal) => literal.format(p),
+            Self::TaggedTemplateExpression(expr) => expr.format(p),
+            Self::Super(sup) => sup.format(p),
+            Self::AwaitExpression(expr) => expr.format(p),
+            Self::ChainExpression(expr) => expr.format(p),
+            Self::NewExpression(expr) => expr.format(p),
+            Self::MetaProperty(expr) => expr.format(p),
+            Self::ClassExpression(expr) => expr.format(p),
+            Self::JSXElement(el) => el.format(p),
+            Self::JSXFragment(fragment) => fragment.format(p),
+            Self::TSAsExpression(expr) => expr.format(p),
+            Self::TSSatisfiesExpression(expr) => expr.format(p),
+            Self::TSTypeAssertion(expr) => expr.format(p),
+            Self::TSNonNullExpression(expr) => expr.format(p),
+            Self::TSInstantiationExpression(expr) => expr.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for IdentifierReference<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, IdentifierReference, { dynamic_text!(p, self.name.as_str()) })
+    }
+}
+
+impl<'a> Format<'a> for IdentifierName<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        dynamic_text!(p, self.name.as_str())
+    }
+}
+
+impl<'a> Format<'a> for BindingIdentifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, BindingIdentifier, { dynamic_text!(p, self.name.as_str()) })
+    }
+}
+
+impl<'a> Format<'a> for LabelIdentifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        dynamic_text!(p, self.name.as_str())
+    }
+}
+
+impl<'a> Format<'a> for BooleanLiteral {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        text!(if self.value { "true" } else { "false" })
+    }
+}
+
+impl<'a> Format<'a> for NullLiteral {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        text!("null")
+    }
+}
+
+impl<'a> Format<'a> for NumericLiteral<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, NumericLiteral, {
+            // See https://github.com/prettier/prettier/blob/3.3.3/src/utils/print-number.js
+            // Perf: the regexes from prettier code above are ported to manual search for performance reasons.
+            let mut string = self.span.source_text(p.source_text).cow_to_ascii_lowercase();
+
+            // Remove unnecessary plus and zeroes from scientific notation.
+            if let Some((head, tail)) = string.split_once('e') {
+                let negative = if tail.starts_with('-') { "-" } else { "" };
+                let trimmed = tail.trim_start_matches(['+', '-']).trim_start_matches('0');
+                if trimmed.starts_with(|c: char| c.is_ascii_digit()) {
+                    string = Cow::Owned(std::format!("{head}e{negative}{trimmed}"));
+                }
+            }
+
+            // Remove unnecessary scientific notation (1e0).
+            if let Some((head, tail)) = string.split_once('e') {
+                if tail.trim_start_matches(['+', '-']).trim_start_matches('0').is_empty() {
+                    string = Cow::Owned(head.to_string());
+                }
+            }
+
+            // Make sure numbers always start with a digit.
+            if string.starts_with('.') {
+                string = Cow::Owned(std::format!("0{string}"));
+            }
+
+            // Remove extraneous trailing decimal zeroes.
+            if let Some((head, tail)) = string.split_once('.') {
+                if let Some((head_e, tail_e)) = tail.split_once('e') {
+                    if !head_e.is_empty() {
+                        let trimmed = head_e.trim_end_matches('0');
+                        if trimmed.is_empty() {
+                            string = Cow::Owned(std::format!("{head}.0e{tail_e}"));
+                        } else {
+                            string = Cow::Owned(std::format!("{head}.{trimmed}e{tail_e}"));
+                        }
+                    }
+                } else if !tail.is_empty() {
+                    let trimmed = tail.trim_end_matches('0');
+                    if trimmed.is_empty() {
+                        string = Cow::Owned(std::format!("{head}.0"));
+                    } else {
+                        string = Cow::Owned(std::format!("{head}.{trimmed}"));
+                    }
+                }
+            }
+
+            // Remove trailing dot.
+            if let Some((head, tail)) = string.split_once('.') {
+                if tail.is_empty() {
+                    string = Cow::Owned(head.to_string());
+                } else if tail.starts_with('e') {
+                    string = Cow::Owned(std::format!("{head}{tail}"));
+                }
+            }
+
+            dynamic_text!(p, &string)
+        })
+    }
+}
+
+impl<'a> Format<'a> for BigIntLiteral<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self.span.source_text(p.source_text).cow_to_ascii_lowercase() {
+            Cow::Borrowed(s) => dynamic_text!(p, s),
+            Cow::Owned(s) => dynamic_text!(p, &s),
+        }
+    }
+}
+
+impl<'a> Format<'a> for RegExpLiteral<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("/"));
+        parts.push(dynamic_text!(p, self.regex.pattern.source_text(p.source_text).as_ref()));
+        parts.push(text!("/"));
+        parts.push(self.regex.flags.format(p));
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for StringLiteral<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, StringLiteral, {
+            let raw = &p.source_text[(self.span.start + 1) as usize..(self.span.end - 1) as usize];
+            // TODO: implement `makeString` from prettier/src/utils/print-string.js
+            dynamic_text!(p, string::print_string(p, raw, p.options.single_quote))
+        })
+    }
+}
+
+impl<'a> Format<'a> for ThisExpression {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        text!("this")
+    }
+}
+
+impl<'a> Format<'a> for MemberExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, MemberExpression, {
+            match self {
+                Self::ComputedMemberExpression(expr) => expr.format(p),
+                Self::StaticMemberExpression(expr) => expr.format(p),
+                Self::PrivateFieldExpression(expr) => expr.format(p),
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for ComputedMemberExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(self.object.format(p));
+        if self.optional {
+            parts.push(text!("?."));
+        }
+        parts.push(text!("["));
+        parts.push(self.expression.format(p));
+        parts.push(text!("]"));
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for StaticMemberExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(self.object.format(p));
+        if self.optional {
+            parts.push(text!("?"));
+        }
+        parts.push(text!("."));
+        parts.push(self.property.format(p));
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for PrivateFieldExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(self.object.format(p));
+        parts.push(if self.optional { text!("?.") } else { text!(".") });
+        parts.push(self.field.format(p));
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for CallExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, CallExpression, {
+            call_expression::print_call_expression(
+                p,
+                &call_expression::CallExpressionLike::CallExpression(self),
+            )
+        })
+    }
+}
+
+impl<'a> Format<'a> for Argument<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            match_expression!(Self) => self.to_expression().format(p),
+            Self::SpreadElement(expr) => expr.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ArrayExpressionElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::SpreadElement(expr) => expr.format(p),
+            match_expression!(Self) => self.to_expression().format(p),
+            Self::Elision(elision) => text!(""),
+        }
+    }
+}
+
+impl<'a> Format<'a> for SpreadElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, SpreadElement, {
+            let argument_doc = self.argument.format(p);
+            array!(p, [text!("..."), argument_doc])
+        })
+    }
+}
+
+impl<'a> Format<'a> for ArrayExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ArrayExpression, {
+            array::print_array(p, &array::Array::ArrayExpression(self))
+        })
+    }
+}
+
+impl<'a> Format<'a> for ObjectExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ObjectExpression, {
+            object::print_object(p, object::ObjectLike::Expression(self))
+        })
+    }
+}
+
+impl<'a> Format<'a> for ObjectPropertyKind<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            ObjectPropertyKind::ObjectProperty(prop) => prop.format(p),
+            ObjectPropertyKind::SpreadProperty(prop) => prop.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ObjectProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ObjectProperty, {
+            if self.method || self.kind == PropertyKind::Get || self.kind == PropertyKind::Set {
+                let mut parts = Vec::new_in(p.allocator);
+                let mut method = self.method;
+                match self.kind {
+                    PropertyKind::Get => {
+                        parts.push(text!("get "));
+                        method = true;
+                    }
+                    PropertyKind::Set => {
+                        parts.push(text!("set "));
+                        method = true;
+                    }
+                    PropertyKind::Init => (),
+                }
+                if method {
+                    if let Expression::FunctionExpression(func_expr) = &self.value {
+                        parts.push(wrap!(p, func_expr, Function, {
+                            function::print_function(
+                                p,
+                                func_expr,
+                                Some(self.key.span().source_text(p.source_text)),
+                            )
+                        }));
+                    }
+                } else {
+                    parts.push(self.key.format(p));
+                    parts.push(text!(": "));
+                    parts.push(self.value.format(p));
+                }
+                return group!(p, parts);
+            }
+
+            if self.shorthand {
+                return self.value.format(p);
+            }
+
+            let left_doc = if self.computed {
+                let key_doc = self.key.format(p);
+                array!(p, [text!("["), key_doc, text!("]")])
+            } else {
+                self.key.format(p)
+            };
+
+            assignment::print_assignment(
+                p,
+                assignment::AssignmentLikeNode::ObjectProperty(self),
+                left_doc,
+                text!(":"),
+                Some(&self.value),
+            )
+        })
+    }
+}
+
+impl<'a> Format<'a> for PropertyKey<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let is_parent_computed = match p.current_kind() {
+            AstKind::MethodDefinition(node) => node.computed,
+            AstKind::PropertyDefinition(node) => node.computed,
+            _ => false,
+        };
+        if is_parent_computed {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("["));
+            let doc = match self {
+                PropertyKey::StaticIdentifier(ident) => ident.format(p),
+                PropertyKey::PrivateIdentifier(ident) => ident.format(p),
+                match_expression!(PropertyKey) => self.to_expression().format(p),
+            };
+            parts.push(doc);
+            parts.push(text!("]"));
+            return array!(p, parts);
+        }
+
+        wrap!(p, self, PropertyKey, {
+            // Perf: Cache the result of `need_quote` to avoid checking it in each PropertyKey
+            let need_quote = p.options.quote_props.consistent()
+                && match p.parent_parent_kind() {
+                    Some(AstKind::ObjectExpression(a)) => a.properties.iter().any(|x| match x {
+                        ObjectPropertyKind::ObjectProperty(p) => {
+                            property::is_property_key_has_quote(&p.key)
+                        }
+                        ObjectPropertyKind::SpreadProperty(_) => false,
+                    }),
+                    Some(AstKind::ClassBody(a)) => a.body.iter().any(|x| match x {
+                        ClassElement::PropertyDefinition(p) => {
+                            property::is_property_key_has_quote(&p.key)
+                        }
+                        _ => false,
+                    }),
+                    _ => false,
+                };
+
+            match self {
+                PropertyKey::StaticIdentifier(ident) => {
+                    if need_quote {
+                        dynamic_text!(
+                            p,
+                            string::print_string(p, &ident.name, p.options.single_quote)
+                        )
+                    } else {
+                        ident.format(p)
+                    }
+                }
+                PropertyKey::PrivateIdentifier(ident) => ident.format(p),
+                PropertyKey::StringLiteral(literal) => {
+                    // This does not pass quotes/objects.js
+                    // because prettier uses the function `isEs5IdentifierName` based on unicode version 3,
+                    // but `is_identifier_name` uses the latest unicode version.
+                    if is_identifier_name(literal.value.as_str())
+                        && (p.options.quote_props.as_needed()
+                            || (p.options.quote_props.consistent()/* && !needsQuoteProps.get(parent) */))
+                    {
+                        dynamic_text!(p, literal.value.as_str())
+                    } else {
+                        dynamic_text!(
+                            p,
+                            string::print_string(p, literal.value.as_str(), p.options.single_quote,)
+                        )
+                    }
+                }
+                PropertyKey::NumericLiteral(literal) => {
+                    if need_quote {
+                        dynamic_text!(
+                            p,
+                            string::print_string(p, &literal.raw_str(), p.options.single_quote)
+                        )
+                    } else {
+                        literal.format(p)
+                    }
+                }
+                PropertyKey::Identifier(ident) => {
+                    let ident_doc = ident.format(p);
+                    array!(p, [text!("["), ident_doc, text!("]")])
+                }
+                match_expression!(PropertyKey) => self.to_expression().format(p),
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for ArrowFunctionExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ArrowFunctionExpression, { arrow_function::print_arrow_function(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for YieldExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, YieldExpression, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("yield"));
+            if self.delegate {
+                parts.push(text!("*"));
+            }
+            if let Some(argument) = &self.argument {
+                parts.push(text!(" "));
+                parts.push(argument.format(p));
+            }
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for UpdateExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, UpdateExpression, {
+            if self.prefix {
+                let argument_doc = self.argument.format(p);
+                array!(p, [text!(self.operator.as_str()), argument_doc])
+            } else {
+                let argument_doc = self.argument.format(p);
+                array!(p, [argument_doc, text!(self.operator.as_str())])
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for UnaryExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, UnaryExpression, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(dynamic_text!(p, self.operator.as_str()));
+            if self.operator.is_keyword() {
+                parts.push(text!(" "));
+            }
+            parts.push(self.argument.format(p));
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for BinaryExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, BinaryExpression, {
+            let doc = binaryish::print_binaryish_expression(
+                p,
+                &self.left,
+                self.operator.into(),
+                &self.right,
+            );
+            if misc::in_parentheses(p.parent_kind(), p.source_text, self.span) {
+                group!(p, [indent!(p, [softline!(), doc]), softline!(),])
+            } else {
+                doc
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for PrivateInExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, PrivateInExpression, {
+            let left_doc = self.left.format(p);
+            let right_doc = self.right.format(p);
+            array!(p, [left_doc, text!(" "), text!(self.operator.as_str()), text!(" "), right_doc])
+        })
+    }
+}
+
+impl<'a> Format<'a> for LogicalExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, LogicalExpression, {
+            let doc = binaryish::print_binaryish_expression(
+                p,
+                &self.left,
+                self.operator.into(),
+                &self.right,
+            );
+
+            if misc::in_parentheses(p.parent_kind(), p.source_text, self.span) {
+                group!(p, [indent!(p, [softline!(), doc]), softline!()])
+            } else {
+                doc
+            }
+        })
+    }
+}
+
+impl<'a> Format<'a> for ConditionalExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ConditionalExpression, { ternary::print_ternary(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for AssignmentExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, AssignmentExpression, { assignment::print_assignment_expression(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTarget<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            match_simple_assignment_target!(Self) => self.to_simple_assignment_target().format(p),
+            match_assignment_target_pattern!(Self) => self.to_assignment_target_pattern().format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for SimpleAssignmentTarget<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::AssignmentTargetIdentifier(ident) => ident.format(p),
+            match_member_expression!(Self) => self.to_member_expression().format(p),
+            Self::TSAsExpression(expr) => expr.format(p),
+            Self::TSSatisfiesExpression(expr) => expr.format(p),
+            Self::TSNonNullExpression(expr) => expr.format(p),
+            Self::TSTypeAssertion(expr) => expr.format(p),
+            Self::TSInstantiationExpression(expr) => expr.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetPattern<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::ArrayAssignmentTarget(target) => target.format(p),
+            Self::ObjectAssignmentTarget(target) => target.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ArrayAssignmentTarget<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        array::print_array(p, &array::Array::ArrayAssignmentTarget(self))
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetMaybeDefault<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            match_assignment_target!(AssignmentTargetMaybeDefault) => {
+                self.to_assignment_target().format(p)
+            }
+            AssignmentTargetMaybeDefault::AssignmentTargetWithDefault(v) => v.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for ObjectAssignmentTarget<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        object::print_object(p, object::ObjectLike::AssignmentTarget(self))
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetWithDefault<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let binding_doc = self.binding.format(p);
+        let init_doc = self.init.format(p);
+        array!(p, [binding_doc, text!(" = "), init_doc])
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::AssignmentTargetPropertyIdentifier(ident) => ident.format(p),
+            Self::AssignmentTargetPropertyProperty(prop) => prop.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetPropertyIdentifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(self.binding.format(p));
+        if let Some(init) = &self.init {
+            parts.push(text!(" = "));
+            parts.push(init.format(p));
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetPropertyProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let name_doc = self.name.format(p);
+        let binding_doc = self.binding.format(p);
+        array!(p, [name_doc, text!(": "), binding_doc])
+    }
+}
+
+impl<'a> Format<'a> for AssignmentTargetRest<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let target_doc = self.target.format(p);
+        array!(p, [text!("..."), target_doc])
+    }
+}
+
+impl<'a> Format<'a> for SequenceExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, SequenceExpression, {
+            let docs =
+                self.expressions.iter().map(|expr| expr.format(p)).collect::<std::vec::Vec<_>>();
+            group!(p, [join!(p, JoinSeparator::CommaLine, docs)])
+        })
+    }
+}
+
+impl<'a> Format<'a> for ParenthesizedExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        unreachable!("Parser preserve_parens option need to be set to false.");
+    }
+}
+
+impl<'a> Format<'a> for ImportExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ImportExpression, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("import"));
+            parts.push(text!("("));
+            let mut indent_parts = Vec::new_in(p.allocator);
+            indent_parts.push(softline!());
+            indent_parts.push(self.source.format(p));
+            if !self.arguments.is_empty() {
+                for arg in &self.arguments {
+                    indent_parts.push(text!(","));
+                    indent_parts.push(line!());
+                    indent_parts.push(arg.format(p));
+                }
+            }
+            parts.push(group!(p, [indent!(p, indent_parts)]));
+            parts.push(softline!());
+            parts.push(text!(")"));
+
+            group!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for TemplateLiteral<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        template_literal::print_template_literal(
+            p,
+            &template_literal::TemplateLiteralPrinter::TemplateLiteral(self),
+        )
+    }
+}
+
+impl<'a> Format<'a> for TemplateElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        // TODO: `replaceEndOfLine`
+        dynamic_text!(p, self.value.raw.as_str())
+    }
+}
+
+impl<'a> Format<'a> for TaggedTemplateExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, TaggedTemplateExpression, {
+            let mut parts = Vec::new_in(p.allocator);
+
+            parts.push(self.tag.format(p));
+
+            if let Some(type_parameters) = &self.type_parameters {
+                parts.push(text!("<"));
+                parts.push(type_parameters.format(p));
+                parts.push(text!(">"));
+            }
+
+            parts.push(self.quasi.format(p));
+
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for Super {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        text!("super")
+    }
+}
+
+impl<'a> Format<'a> for AwaitExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, AwaitExpression, {
+            let mut parts = Vec::new_in(p.allocator);
+            parts.push(text!("await "));
+            parts.push(self.argument.format(p));
+            array!(p, parts)
+        })
+    }
+}
+
+impl<'a> Format<'a> for ChainExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ChainExpression, { self.expression.format(p) })
+    }
+}
+
+impl<'a> Format<'a> for ChainElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            Self::CallExpression(expr) => expr.format(p),
+            Self::TSNonNullExpression(expr) => expr.format(p),
+            match_member_expression!(Self) => self.to_member_expression().format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for NewExpression<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, NewExpression, {
+            call_expression::print_call_expression(
+                p,
+                &call_expression::CallExpressionLike::NewExpression(self),
+            )
+        })
+    }
+}
+
+impl<'a> Format<'a> for MetaProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let meta_doc = self.meta.format(p);
+        let property_doc = self.property.format(p);
+        array!(p, [meta_doc, text!("."), property_doc])
+    }
+}
+
+impl<'a> Format<'a> for Class<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, Class, { class::print_class(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for ClassBody<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ClassBody, { class::print_class_body(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for ClassElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        match self {
+            ClassElement::StaticBlock(c) => c.format(p),
+            ClassElement::MethodDefinition(c) => c.format(p),
+            ClassElement::PropertyDefinition(c) => c.format(p),
+            ClassElement::AccessorProperty(c) => c.format(p),
+            ClassElement::TSIndexSignature(c) => c.format(p),
+        }
+    }
+}
+
+impl<'a> Format<'a> for StaticBlock<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, StaticBlock, {
+            let block_doc = block::print_block(p, &self.body, None);
+            array!(p, [text!("static "), block_doc])
+        })
+    }
+}
+
+impl<'a> Format<'a> for MethodDefinition<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, MethodDefinition, { function::print_method(p, self) })
+    }
+}
+
+impl<'a> Format<'a> for PropertyDefinition<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, PropertyDefinition, {
+            class::print_class_property(p, &class::ClassMemberish::PropertyDefinition(self))
+        })
+    }
+}
+
+impl<'a> Format<'a> for AccessorProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        class::print_class_property(p, &class::ClassMemberish::AccessorProperty(self))
+    }
+}
+
+impl<'a> Format<'a> for PrivateIdentifier<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(text!("#"));
+        parts.push(dynamic_text!(p, self.name.as_str()));
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for BindingPattern<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut parts = Vec::new_in(p.allocator);
+        parts.push(match self.kind {
+            BindingPatternKind::BindingIdentifier(ref ident) => ident.format(p),
+            BindingPatternKind::ObjectPattern(ref pattern) => pattern.format(p),
+            BindingPatternKind::ArrayPattern(ref pattern) => pattern.format(p),
+            BindingPatternKind::AssignmentPattern(ref pattern) => pattern.format(p),
+        });
+
+        if self.optional {
+            parts.push(text!("?"));
+        }
+
+        if let Some(typ) = &self.type_annotation {
+            let type_annotation_doc = typ.type_annotation.format(p);
+            parts.push(array!(p, [text!(": "), type_annotation_doc]));
+        }
+        array!(p, parts)
+    }
+}
+
+impl<'a> Format<'a> for ObjectPattern<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ObjectPattern, {
+            object::print_object(p, object::ObjectLike::Pattern(self))
+        })
+    }
+}
+
+impl<'a> Format<'a> for BindingProperty<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        if self.shorthand {
+            self.value.format(p)
+        } else {
+            let key_doc = self.key.format(p);
+            let value_doc = self.value.format(p);
+            group!(p, [key_doc, text!(": "), value_doc])
+        }
+    }
+}
+
+impl<'a> Format<'a> for BindingRestElement<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let argument_doc = self.argument.format(p);
+        array!(p, [text!("..."), argument_doc])
+    }
+}
+
+impl<'a> Format<'a> for ArrayPattern<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, ArrayPattern, { array::print_array(p, &array::Array::ArrayPattern(self)) })
+    }
+}
+
+impl<'a> Format<'a> for AssignmentPattern<'a> {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        wrap!(p, self, AssignmentPattern, {
+            let left_doc = self.left.format(p);
+            let right_doc = self.right.format(p);
+            array!(p, [left_doc, text!(" = "), right_doc])
+        })
+    }
+}
+
+impl<'a> Format<'a> for RegExpFlags {
+    fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
+        let mut string = std::vec::Vec::with_capacity(self.iter().count());
+        if self.contains(Self::D) {
+            string.push('d');
+        }
+        if self.contains(Self::G) {
+            string.push('g');
+        }
+        if self.contains(Self::I) {
+            string.push('i');
+        }
+        if self.contains(Self::M) {
+            string.push('m');
+        }
+        if self.contains(Self::S) {
+            string.push('s');
+        }
+        if self.contains(Self::U) {
+            string.push('u');
+        }
+        if self.contains(Self::V) {
+            string.push('v');
+        }
+        if self.contains(Self::Y) {
+            string.push('y');
+        }
+        let sorted = string.iter().collect::<String>();
+        dynamic_text!(p, &sorted)
+    }
+}
