+++ crates/oxc_prettier/src/format/print/call_expression.rs
@@ -2,9 +2,11 @@ use oxc_allocator::Vec;
 use oxc_ast::ast::*;
 use oxc_span::{GetSpan, Span};
 
-use crate::{format::call_arguments::print_call_arguments, group, ir::Doc, text, Format, Prettier};
+use crate::{
+    format::print::call_arguments::print_call_arguments, group, ir::Doc, text, Format, Prettier,
+};
 
-pub(super) enum CallExpressionLike<'a, 'b> {
+pub enum CallExpressionLike<'a, 'b> {
     CallExpression(&'b CallExpression<'a>),
     NewExpression(&'b NewExpression<'a>),
 }
@@ -35,9 +37,7 @@ impl<'a> CallExpressionLike<'a, '_> {
         }
     }
 
-    fn type_parameters(
-        &self,
-    ) -> Option<&oxc_allocator::Box<'a, TSTypeParameterInstantiation<'a>>> {
+    fn type_parameters(&self) -> Option<&oxc_allocator::Box<'a, TSTypeParameterInstantiation<'a>>> {
         match self {
             CallExpressionLike::CallExpression(call) => call.type_parameters.as_ref(),
             CallExpressionLike::NewExpression(new) => new.type_parameters.as_ref(),
@@ -54,7 +54,7 @@ impl GetSpan for CallExpressionLike<'_, '_> {
     }
 }
 
-pub(super) fn print_call_expression<'a>(
+pub fn print_call_expression<'a>(
     p: &mut Prettier<'a>,
     expression: &CallExpressionLike<'a, '_>,
 ) -> Doc<'a> {
@@ -80,7 +80,7 @@ pub(super) fn print_call_expression<'a>(
 }
 
 /// <https://github.com/prettier/prettier/blob/7aecca5d6473d73f562ca3af874831315f8f2581/src/language-js/print/call-expression.js#L93-L116>
-pub(super) fn is_commons_js_or_amd_call<'a>(
+pub fn is_commons_js_or_amd_call<'a>(
     callee: &Expression<'a>,
     arguments: &Vec<'a, Argument<'a>>,
 ) -> bool {
