+++ crates/oxc_prettier/src/format/print/mod.rs
@@ -1,0 +1,18 @@
+pub mod array;
+pub mod arrow_function;
+pub mod assignment;
+pub mod binaryish;
+pub mod block;
+pub mod call_arguments;
+pub mod call_expression;
+pub mod class;
+pub mod function;
+pub mod function_parameters;
+pub mod misc;
+pub mod module;
+pub mod object;
+pub mod property;
+pub mod statement;
+pub mod string;
+pub mod template_literal;
+pub mod ternary;
