+++ crates/oxc_prettier/src/format/print/string.rs
@@ -55,11 +55,7 @@ fn make_string<'a>(p: &Prettier<'a>, raw_text: &str, enclosing_quote: char) -> S
     result
 }
 
-pub(super) fn print_string<'a>(
-    p: &Prettier<'a>,
-    raw_text: &str,
-    prefer_single_quote: bool,
-) -> &'a str {
+pub fn print_string<'a>(p: &Prettier<'a>, raw_text: &str, prefer_single_quote: bool) -> &'a str {
     let enclosing_quote = get_preferred_quote(raw_text, prefer_single_quote);
     make_string(p, raw_text, enclosing_quote).into_bump_str()
 }
